Temporary fixing Quiz Page
