+++ frontend/src/services/gptServices.js
@@ -10,6 +10,7 @@ export async function gptServices(payload) {
         model: "gpt-3.5-turbo",
         messages: payload,
         temperature: 0.7,
+        max_tokens: 1000,
       },
       {
         headers: {
@@ -18,10 +19,9 @@ export async function gptServices(payload) {
         },
       }
     );
-    console.log("[test]: ",response)
-    const answer = response.data.choices[0]?.message?.content || "";
+    const answer = response.data.choices[0]?.message?.content.trim() || "";
     return answer;
   } catch (err) {
-    console.error("Error in fetching or generating response:", err);
+    console.error("Error in generating response in gptService:", err.message);
   }
 }
